# docs/api/api-docs.yaml
openapi: 3.0.1
info:
  title: QProv API
  contact:
    name: GitHub Repository
    url: https://github.com/UST-QuAntiL/QProv
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.3.0
servers:
- url: http://localhost:5020/qprov
  description: Generated server url
paths:
  /virtual-machines/{virtualMachineId}/characteristics:
    get:
      tags:
      - virtual-machine
      description: "Retrieve the calibration characteristics from the given gate.\
        \ By using the latest parameter only the latest data is retrieved, otherwise\
        \ all available data."
      operationId: getHardwareCharacteristics
      parameters:
      - name: virtualMachineId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: latest
        in: query
        required: true
        schema:
          type: boolean
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/CollectionModelEntityModelHardwareCharacteristicsDto'
        "404":
          description: No characteristics for this Gate available.
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/CollectionModelEntityModelHardwareCharacteristicsDto'
    put:
      tags:
      - virtual-machine
      description: "Create a new VirtualMachine and return the link which can then\
        \ be used to retrieve, update, and delete it."
      operationId: addHardwareCharacteristics
      parameters:
      - name: virtualMachineId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HardwareCharacteristicsDto'
        required: true
      responses:
        "201":
          description: Created
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/EntityModelVirtualMachineDto'
...
  /providers/{providerId}/qpus/{qpuId}/qubits/{qubitId}/gates:
    get:
      tags:
      - provider
      description: Retrieve all Gates that can be executed on the Qubit.
      operationId: getGates
      parameters:
      - name: providerId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: qpuId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: qubitId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/CollectionModelEntityModelGateDto'
        "404":
          description: Qubit or QPU not available.
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/CollectionModelEntityModelGateDto'
  /providers/{providerId}/qpus/{qpuId}/qubits/{qubitId}/gates/{gateId}:
    get:
      tags:
      - provider
      description: Retrieve a specific Qubit and its basic properties.
      operationId: getGate
      parameters:
      - name: providerId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: qpuId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: qubitId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: gateId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/EntityModelGateDto'
        "404":
          description: Not Found. Gate with given ID doesn't exist.
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/EntityModelGateDto'
  /providers/{providerId}/qpus/{qpuId}/qubits/{qubitId}/gates/{gateId}/characteristics:
    get:
      tags:
      - provider
      description: "Retrieve the calibration characteristics from the given gate.\
        \ By using the latest parameter only the latest data is retrieved, otherwise\
        \ all available data."
      operationId: getGateCharacterisitcs
      parameters:
      - name: providerId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: qpuId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: qubitId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: gateId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: latest
        in: query
        required: true
        schema:
          type: boolean
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/CollectionModelEntityModelGateCharacteristicsDto'
        "404":
          description: No characteristics for this Gate available.
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/CollectionModelEntityModelGateCharacteristicsDto'
  /providers/{providerId}/qpus/{qpuId}/qubits/{qubitId}/characteristics:
    get:
      tags:
      - provider
      description: "Retrieve the calibration characteristics from the given qubit.\
        \ By using the latest parameter only the latest data is retrieved, otherwise\
        \ all available data."
      operationId: getQubitCharacterisitcs
      parameters:
      - name: providerId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: qpuId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: qubitId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: latest
        in: query
        required: true
        schema:
          type: boolean
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/CollectionModelEntityModelQubitCharacteristicsDto'
        "404":
          description: No characteristics for this Qubit available.
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/CollectionModelEntityModelQubitCharacteristicsDto'
  /providers/{providerId}/qpus/{qpuId}/aggregated-data:
    get:
      tags:
      - provider
      operationId: getLinksToAggregatedData
      parameters:
      - name: providerId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: qpuId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/RepresentationModelObject'
        "404":
          description: Provider or QPU not found.
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/RepresentationModelObject'
  /providers/{providerId}/qpus/{qpuId}/aggregated-data/calibration-matrix:
    get:
      tags:
      - provider
      operationId: getCalibrationMatrix
      parameters:
      - name: providerId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: qpuId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: latest
        in: query
        required: true
        schema:
          type: boolean
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/CollectionModelEntityModelCalibrationMatrixDto'
        "404":
          description: Provider or QPU not found or no calibration matrix available
            for this QPU.
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/CollectionModelEntityModelCalibrationMatrixDto'
...
components:
  schemas:
    HardwareCharacteristicsDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        recordingTime:
          type: string
          format: date-time
        cpuUsage:
          type: number
          format: float
        clockSpeed:
          type: number
          format: float
        ramUsage:
          type: number
          format: float
        diskUsage:
          type: number
          format: float
    EntityModelVirtualMachineDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        name:
          type: string
        cpu:
          type: string
        cpuCores:
          type: integer
          format: int32
        ramSize:
          type: integer
          format: int32
        diskSize:
          type: integer
          format: int32
        _links:
          $ref: '#/components/schemas/Links'
    Links:
      type: object
      additionalProperties:
        $ref: '#/components/schemas/Link'
...
    CollectionModelEntityModelHardwareCharacteristicsDto:
      type: object
      properties:
        _embedded:
          type: object
          properties:
            hardwareCharacteristicsDtoes:
              type: array
              items:
                $ref: '#/components/schemas/EntityModelHardwareCharacteristicsDto'
        _links:
          $ref: '#/components/schemas/Links'
    EntityModelHardwareCharacteristicsDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        recordingTime:
          type: string
          format: date-time
        cpuUsage:
          type: number
          format: float
        clockSpeed:
          type: number
          format: float
        ramUsage:
          type: number
          format: float
        diskUsage:
          type: number
          format: float
        _links:
          $ref: '#/components/schemas/Links'
...
    CollectionModelEntityModelGateDto:
      type: object
      properties:
        _embedded:
          type: object
          properties:
            gateDtoes:
              type: array
              items:
                $ref: '#/components/schemas/EntityModelGateDto'
        _links:
          $ref: '#/components/schemas/Links'
    EntityModelGateDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        multiQubitGate:
          type: boolean
        operatingQubits:
          type: array
          items:
            type: string
            format: uuid
        _links:
          $ref: '#/components/schemas/Links'
    CollectionModelEntityModelGateCharacteristicsDto:
      type: object
      properties:
        _embedded:
          type: object
          properties:
            gateCharacteristicsDtoes:
              type: array
              items:
                $ref: '#/components/schemas/EntityModelGateCharacteristicsDto'
        _links:
          $ref: '#/components/schemas/Links'
    EntityModelGateCharacteristicsDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        calibrationTime:
          type: string
          format: date-time
        gateTime:
          type: number
        gateErrorRate:
          type: number
        _links:
          $ref: '#/components/schemas/Links'
    CollectionModelEntityModelQubitCharacteristicsDto:
      type: object
      properties:
        _embedded:
          type: object
          properties:
            qubitCharacteristicsDtoes:
              type: array
              items:
                $ref: '#/components/schemas/EntityModelQubitCharacteristicsDto'
        _links:
          $ref: '#/components/schemas/Links'
    EntityModelQubitCharacteristicsDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        calibrationTime:
          type: string
          format: date-time
        t1Time:
          type: number
        t2Time:
          type: number
        readoutError:
          type: number
        _links:
          $ref: '#/components/schemas/Links'
    RepresentationModelObject:
      type: object
      properties:
        _links:
          $ref: '#/components/schemas/Links'
    CollectionModelEntityModelCalibrationMatrixDto:
      type: object
      properties:
        _embedded:
          type: object
          properties:
            calibrationMatrixDtoes:
              type: array
              items:
                $ref: '#/components/schemas/EntityModelCalibrationMatrixDto'
        _links:
          $ref: '#/components/schemas/Links'
    EntityModelCalibrationMatrixDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        calibrationTime:
          type: string
          format: date-time
        calibrationMatrix:
          type: array
          items:
            type: array
            items:
              type: number
              format: double
        _links:
          $ref: '#/components/schemas/Links'
...
    Link:
      type: object
      properties:
        href:
          type: string
        hreflang:
          type: string
        title:
          type: string
        type:
          type: string
        deprecation:
          type: string
        profile:
          type: string
        name:
          type: string
        templated:
          type: boolean